% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.shiny-package.R
\docType{package}
\name{nba.shiny-package}
\alias{nba.shiny}
\alias{nba.shiny-package}
\title{nba.shiny: NBA Shiny}
\description{
A Shiny application for NBA Fantasy
}
\author{
\strong{Maintainer}: Oliver Eaton \email{eat_fred@proton.me}

Authors:
\itemize{
  \item Oliver Eaton \email{eat_fred@proton.me}
}

}
